Fix: Ban/Unban Vendor refactor
